statemachine/requirements.txt
rpi-lgpio==0.6
requests==2.32.3
websocket-client==1.8.0 
python-dotenv==1.0.1